api_service/requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Flask-Limiter==3.9.2
PyJWT==2.11.0
bcrypt==5.0.0
gunicorn==25.1.0
uvicorn==0.41.0
asgiref==3.11.1

aiohttp==3.13.3
redis==7.2.0
amqp==5.3.1
billiard==4.2.4
kombu==5.6.2

APScheduler==3.11.2
croniter==6.0.0

concurrent-log-handler==0.9.29
requests==2.32.5
python-dotenv==1.2.1
PyYAML==6.0.3
pytz==2025.2
tzdata==2025.3

PlexAPI==4.18.0
simple-justwatch-python-api==0.16
xmltodict==1.0.4

mysql-connector-python==9.6.0
psycopg2-binary==2.9.11

openai==2.24.0